add access control to mutations

+++ backend/index.js
@@ -12,6 +12,7 @@ const tmc = require('./TMCAuthentication')
 const jwt = require('jsonwebtoken')
 const { AuthenticationError } = require('apollo-server-core')
 const { authenticate } = require('./middleware/authentication')
+const { checkAccess } = require('./accessControl')
 
 const resolvers = {
   Query: {
@@ -89,6 +90,7 @@ const resolvers = {
     },
 
     async deleteCourseAsCoursePrerequisite(root, args, context) {
+      checkAccess(context, { allowUser: true })
       await context.prisma.updateCourse({
         where: { id: args.id },
         data: {
@@ -102,6 +104,7 @@ const resolvers = {
       })
     },
     async addCourseAsCoursePrerequisite(root, args, context) {
+      checkAccess(context, { allowUser: true })
       await context.prisma.updateCourse({
         where: { id: args.id },
         data: {
@@ -115,6 +118,7 @@ const resolvers = {
       })
     },
     createResourceWithURLs(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return context.prisma.createResource({
         name: args.name,
         description: args.desc,
@@ -129,12 +133,14 @@ const resolvers = {
       })
     },
     createURL(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return context.prisma.createURL({
         address: args.address,
         resource: { connect: { id: args.resource_id } }
       })
     },
     createResource(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return context.prisma.createResource({
         name: args.name,
         description: args.desc,
@@ -142,16 +148,19 @@ const resolvers = {
       })
     },
     createCourse(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return context.prisma.createCourse({
         name: args.name
       })
     },
     deleteCourse(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return context.prisma.deleteCourse({
         id: args.id
       })
     },
     createConcept(root, args, context) {
+      checkAccess(context, { allowUser: true })
       const concept = args.desc !== undefined
         ? args.official !== undefined
           ? { name: args.name, description: args.desc, official: args.official }
@@ -166,6 +175,7 @@ const resolvers = {
       })
     },
     createConceptAsPrerequisite(root, args, context) {
+      checkAccess(context, { allowUser: true })
       const concept = args.desc !== undefined
         ? args.official !== undefined
           ? { name: args.name, description: args.desc, official: args.official }
@@ -180,6 +190,7 @@ const resolvers = {
       })
     },
     createConceptAsLearningObjective(root, args, context) {
+      checkAccess(context, { allowUser: true })
       const concept = args.desc !== undefined
         ? args.official !== undefined
           ? { name: args.name, description: args.desc, official: args.official }
@@ -194,6 +205,7 @@ const resolvers = {
       })
     },
     updateConcept(root, args, context) {
+      checkAccess(context, { allowUser: true })
       let data = {}
       if (args.name !== undefined) {
         data.name = args.name
@@ -207,12 +219,14 @@ const resolvers = {
       })
     },
     updateCourse(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return context.prisma.updateCourse({
         where: { id: args.id },
         data: { name: args.name }
       })
     },
     async createConceptAndLinkTo(root, args, context) {
+      checkAccess(context, { allowUser: true })
       const concept = args.desc !== undefined
         ? args.official !== undefined
           ? { name: args.name, description: args.desc, official: args.official }
@@ -235,6 +249,7 @@ const resolvers = {
         })
     },
     async deleteConcept(root, args, context) {
+      checkAccess(context, { allowUser: true })
       await context.prisma.deleteManyLinks({
 
         OR: [
@@ -257,6 +272,7 @@ const resolvers = {
       })
     },
     createLink(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return args.official !== undefined
         ? context.prisma.createLink({
           to: {
@@ -277,6 +293,7 @@ const resolvers = {
         })
     },
     deleteLink(root, args, context) {
+      checkAccess(context, { allowUser: true })
       return context.prisma.deleteLink({
         id: args.id
       })
